Don't recognize magic encoding comment as class documentation.
The exact spec for this comment is unclear from the Ruby source. There
is exactly one test for a valid comment, and `parse.y` is ... obtuse.
The regex used here was found in `lib/erb.rb` and copied verbatim, with
the exception of a redundant regex to exclude -*- markers, which AFAICT
is entirely unnecessary.

Fixes #21.