Ignore invalid UTF-8 input when parsing lines.
This continues the hack that cane assumes UTF-8 input, which may not
actually be the case but still seems like a decent option since I don't
have good examples of this not being the case.

This should not add any performance overhead, since input is only
re-encoded if any comparison fails on it. I have not benchmarked whether
this actually a big deal or not, but should.

Fixes #22.